seperated Survey into its own component, removed unused dependencies in Profilecheckin/Survey.tsx, solved a warning where useEffect in Survey did not have handleSurveyOpen as its dependency, solved this by including handleSurveyOpen as a dependency and wrapping handleSurveyOpen with useCallback to make sure that a new handleSurveyOpen is called only when there is a new version of user (dispact/onOpen shouldn't change)
